requirements.txt
# Universal Credit Dashboard Requirements

# Core dependencies
dash>=2.14.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0

# Data processing
pyarrow>=12.0.0
openpyxl>=3.1.0

# API and web requests
requests>=2.31.0
python-dotenv>=1.0.0

# Date/time handling
python-dateutil>=2.8.0

# Additional visualization
plotly-express>=0.4.1

# Development dependencies (optional)
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Geographic data processing (for future enhancements)
geopandas>=0.13.0
folium>=0.14.0